Split registry logic in separate files
Placed all functions related to generating registry events into a separate file.
Created registry.h in order to expose public functions.
